deleted auth module and moved all file to gateway, gateway is our auth, and added config for docker, might be changed
