--- auth/Dockerfile
@@ -1,5 +1,0 @@
-FROM openjdk:8
-
-ADD target/admin.jar .
-
-CMD java -jar admin.jar  --spring.profiles.active=docker
