--- auth/src/main/kotlin/no/ecm/auth/AuthApplicationConfig.kt
@@ -1,34 +1,0 @@
-package no.ecm.auth
-
-import org.springframework.boot.autoconfigure.domain.EntityScan
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import springfox.documentation.builders.ApiInfoBuilder
-import springfox.documentation.builders.PathSelectors
-import springfox.documentation.service.ApiInfo
-import springfox.documentation.spi.DocumentationType
-import springfox.documentation.spring.web.plugins.Docket
-import springfox.documentation.swagger2.annotations.EnableSwagger2
-
-@Configuration
-@EnableSwagger2
-@EntityScan(basePackages = ["no.ecm.auth"])
-class GatewayApplicationConfig {
-
-    @Bean
-    fun swaggerApi(): Docket {
-        return Docket(DocumentationType.SWAGGER_2)
-                .apiInfo(apiInfo())
-                .select()
-                .paths(PathSelectors.any())
-                .build()
-    }
-
-    private fun apiInfo(): ApiInfo {
-        return ApiInfoBuilder()
-                .title("API for auth entity")
-                .description("MicroService that contains auth entity enrepository")
-                .version("1.0")
-                .build()
-    }
-}
