--- auth/src/main/resources/application.yml
@@ -1,24 +1,0 @@
-eureka:
-  client:
-    healthcheck:
-      enabled: true
-    serviceUrl:
-      defaultZone: http://localhost:8761/eureka/
-
-
-spring:
-  application:
-    name: auth-server
-  datasource:
-    url: "jdbc:h2:mem:"
-    driver-class-name: "org.h2.Driver"
-  jpa:
-    database: "h2"
-    hibernate:
-      ddl-auto: "create-drop"
-
-server:
-  port: 8086
-
-cinemaService: http://localhost:8085
-movieService: http://localhost:8083
