--- auth/src/main/resources/ehcache.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ehcache>
-
-    <defaultCache
-            maxElementsInMemory="100"
-            eternal="false"
-            timeToIdleSeconds="120"
-            timeToLiveSeconds="200" />
-
-    <cache
-        name="httpClient"
-        maxElementsInMemory="2000"
-        timeToLiveSeconds="600"
-        eternal="false"
-        overflowToDisk="false" />
-</ehcache>
