--- auth/src/test/resources/application.yml
@@ -1,5 +1,0 @@
-server:
-  port: 7085
-
-cinemaService: http://localhost:8085
-movieService: http://localhost:8083
