+++ cinema/src/main/resources/application.yml
@@ -18,4 +18,4 @@ spring:
       ddl-auto: "create-drop"
 
 server:
-  port: 8085
+  port: 8086
