+++ creditcard/src/main/resources/application.yml
@@ -24,6 +24,4 @@ spring:
       ddl-auto: "create-drop"
 
 server:
-  port: 8084
-
-authService: http://localhost:8086
+  port: 8085
