+++ creditcard/src/test/resources/application-test.yml
@@ -1,11 +1,2 @@
-spring:
-  datasource:
-    url: "jdbc:h2:mem:"
-    driver-class-name: "org.h2.Driver"
-  jpa:
-    database: "h2"
-    hibernate:
-      ddl-auto: "create-drop"
-
-
-authService: http://localhost:8086
+server:
+  port: 7085
