+++ gateway/src/main/kotlin/no/ecm/gateway/GatewayApplication.kt
@@ -2,8 +2,10 @@ package no.ecm.gateway
 
 import org.springframework.boot.SpringApplication
 import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient
 
 @SpringBootApplication(scanBasePackages = ["no.ecm.gateway"])
+@EnableEurekaClient
 class GatewayApplication
 
 fun main(args: Array<String>) {
