+++ gateway/src/main/kotlin/no/ecm/gateway/controller/AuthController.kt
@@ -1,4 +1,4 @@
-package no.ecm.order.controller
+package no.ecm.gateway.controller
 
 import io.swagger.annotations.Api
 import org.springframework.beans.factory.annotation.Value
@@ -8,18 +8,21 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
-@Api(value = "/order", description = "API for order entity")
+@Api(value = "/auth", description = "API for authentication")
 @RequestMapping(
-        path = ["/order"],
+        path = ["/auth"],
         produces = [MediaType.APPLICATION_JSON_UTF8_VALUE])
 @RestController
-class OrderController {
+class AuthController {
+
+    @Value("\${cinemaService}")
+    private lateinit var cinemaHost : String
 
     @Value("\${movieService}")
     private lateinit var movieHost : String
 
     @GetMapping
     fun get(): ResponseEntity<String>? {
-        return ResponseEntity.ok("Invoice")
+        return ResponseEntity.ok("Auth")
     }
 }
