+++ gateway/src/main/kotlin/no/ecm/gateway/controller/RoleType.kt
@@ -1,4 +1,4 @@
-package no.ecm.auth.controller
+package no.ecm.gateway.controller
 
 enum class RoleType {
