+++ gateway/src/main/kotlin/no/ecm/gateway/model/entity/Auth.kt
@@ -1,7 +1,10 @@
-package no.ecm.auth.model.entity
+package no.ecm.gateway.model.entity
 
-import no.ecm.auth.controller.RoleType
-import javax.persistence.*
+import no.ecm.gateway.controller.RoleType
+import javax.persistence.Entity
+import javax.persistence.EnumType
+import javax.persistence.Enumerated
+import javax.persistence.Id
 import javax.validation.constraints.NotBlank
 import javax.validation.constraints.NotNull
 import javax.validation.constraints.Size
