+++ gateway/src/main/resources/application-docker.yml
@@ -1,0 +1,14 @@
+spring:
+  application:
+    name: gateway
+
+eureka:
+  instance:
+    preferIpAddress: true
+  client:
+    healthcheck:
+      enabled: true
+    serviceUrl:
+      defaultZone: http://eureka:8761/eureka/
+server:
+  port: 8080
