+++ gateway/src/main/resources/application.yml
@@ -5,4 +5,7 @@ eureka:
     healthcheck:
       enabled: true
     serviceUrl:
-      defaultZone: http://eureka:8761/eureka/
+      defaultZone: http://localhost:8761/eureka/
+
+server:
+  port: 8084
