+++ movie/src/test/resources/application-test.yml
@@ -1,4 +1,4 @@
 server:
   port: 7083
 
-cinemaService: http://localhost:8085
+cinemaService: http://localhost:8086
