+++ pom.xml
@@ -9,7 +9,6 @@
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>auth</module>
         <module>e2etest</module>
         <module>eureka</module>
         <module>gateway</module>
