+++ user/src/main/kotlin/no/ecm/user/controller/UserController.kt
@@ -13,9 +13,6 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 class UserController {
 
-    @Value("\${authService}")
-    private lateinit var authHost : String
-
     @GetMapping
     fun get(): ResponseEntity<String>? {
         return ResponseEntity.ok("User")
