+++ user/src/main/resources/application.yml
@@ -25,6 +25,3 @@ spring:
 
 server:
   port: 8081
-
-
-authService: http://localhost:8086
