+++ user/src/test/resources/application-test.yml
@@ -1,4 +1,2 @@
 server:
   port: 7081
-
-authService: http://localhost:8086
